requirements.txt
selenium~=3.12.0
Pillow>=7.1.0
beautifulsoup4~=4.6.3
PyAutoGUI~=0.9.52
PyGetWindow~=0.0.9
ibm-watson~=4.7.1
python-dotenv~=0.14.0
pandas~=1.1.3
splunk-sdk~=1.6.14
PyGetWindow~=0.0.9
requests~=2.24.0
PyMongo~=3.11.0
pymongo[srv]~=3.11.0